refactor type changes
