+++ src/app/profile/page.tsx
@@ -7,7 +7,7 @@ import {
   insertProfile,
   updateProfile,
 } from '../../api/supabase/queries/profiles';
-import { Profile } from '../../types/schemaTypes';
+import { Profile } from '../../types/schema';
 
 import styles from '../page.module.css';
