+++ src/types/schema.ts
@@ -2,18 +2,21 @@ import { UUID } from 'crypto';
 
 export interface CaseListing {
   id: UUID;
+  title: string;
   summary: string;
-  languages: string[];
   country: string;
-  legal_server_id: number;
-  client_initials: string;
-  time_to_complete: string;
-  is_remote: boolean;
   client_location: string;
-  program: string;
-  upcoming_hearing_date: string;
+  legal_server_id: number;
+  hours_per_month: number;
+  num_months: number;
+  languages: string[];
+  is_remote: boolean;
+  in_court: boolean;
+  needs_attorney: boolean;
   needs_interpreter: boolean;
-  interest_ids: UUID[];
+  upcoming_date: string;
+  relief_sought: string;
+  active: boolean;
 }
 
 export interface LimitedAssistance {
@@ -34,26 +37,29 @@ export interface TranslationRequest {
 }
 
 export interface Profile {
+  user_id: UUID;
   first_name: string;
   last_name: string;
-  user_id: UUID;
+  location: string;
   roles: string[];
-  languages: string[];
-  accreditations: string[];
-  hours_per_week: number;
+  hours_per_month: number;
   immigration_law_experience: string;
+  languages: string[];
   bar_number: string;
   start_date: string;
+  availability_description: string;
+  eoir_registered: boolean;
   interest_ids: UUID[];
 }
 
 export interface Interest {
   id: UUID;
   listing_id: UUID;
-  listing_type: string;
   user_id: UUID;
+  listing_type: number;
   form_response: {
-    interestType: string[];
+    interestType: string;
     interestReason: string;
+    start_date: string;
   };
 }
